Add product image

+++ src/main/java/com/ratnakar/ecom/repository/ProductRepository.java
@@ -3,6 +3,6 @@ package com.ratnakar.ecom.repository;
 import com.ratnakar.ecom.model.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Products, Integer> {
+public interface ProductRepository extends JpaRepository<Products, Long> {
 
 }
